Cast `TermFeatureFlags` used as `printf()` argument to `unsigned int`
The type of enums is implementation defined, whereas the `%x` format
specifier always requires an `unsigned int` argument.

See also: commit 424f3709f1bcc4540cb672288883c1581680e1cc

+++ src/terminal/feature.c
@@ -170,7 +170,7 @@ UNITTEST {
             if (count > 1) {
                 BUG (
                     "TermEntry '%s' has %u mutually exclusive flags: 0x%x",
-                    name, count, flag_union
+                    name, count, (unsigned int)flag_union
                 );
             }
         }
